# packages/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 临床中心台 SiteDesk API
  version: 0.2.0
  description: |-
    **本文件是产物。** 契约的唯一定义源是 `packages/contracts/src` 下的 zod schema。

    分层：L1 资源层（CRUD 与列表）· L2 命令层（领域动作，返回 sideEffects）· L3 投影层（只读聚合，永不返回明细）。

    带 `x-gated-by` 的字段受列维度权限管辖：**无权限时该字段从响应中消失，不是返回 null**。

    行范围之外的资源一律返回 404 而非 403 —— 403 等于确认「它存在，只是你不能碰」，这个确认本身就是泄漏。
servers:
  - url: https://api.sitedesk.example
    description: 生产
...
paths:
...
  /v1/studies:
    get:
      operationId: listStudies
      tags:
        - site
      summary: 项目列表
      description: 只返回本行范围内有中心的项目 —— 范围之外的项目不存在，不是「无权访问」。
      x-layer: L1
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 512
        - name: limit
          in: query
          required: false
          schema:
            default: 50
            type: integer
            minimum: 1
            maximum: 200
        - name: q
          in: query
          required: false
          schema:
            type: string
            maxLength: 64
      responses:
        '200':
          description: 项目列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListStudiesResponse'
        '401':
          description: 未认证
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/unauthenticated
                title: 未认证
                status: 401
                code: unauthenticated
        '403':
          description: 无此动作权限
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/forbidden-action
                title: 无此动作权限
                status: 403
                code: forbidden-action
        '404':
          description: 资源不存在
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/not-found
                title: 资源不存在
                status: 404
                code: not-found
        '422':
          description: 请求校验未通过
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/validation-failed
                title: 请求校验未通过
                status: 422
                code: validation-failed
        '429':
          description: 请求过于频繁
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/rate-limited
                title: 请求过于频繁
                status: 429
                code: rate-limited
        '500':
          description: 服务内部错误
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/internal
                title: 服务内部错误
                status: 500
                code: internal
  /v1/study-sites:
    get:
      operationId: listStudySites
      tags:
        - site
      summary: 中心台账
      description: |-
        行范围由登录身份推导：CRA 只看被指派的、PM 看本组承接项目下的、机构办只看本院、PI 只看自己担任研究者的。
        **调用方不能通过传参扩大范围**，只能在范围内再收窄。
      x-layer: L1
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 512
        - name: limit
          in: query
          required: false
          schema:
            default: 50
            type: integer
            minimum: 1
            maximum: 200
        - name: studyId
          in: query
          required: false
          schema:
            type: string
            format: uuid
            pattern: >-
              ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
        - name: state
          in: query
          required: false
          schema:
            type: array
            items:
              $ref: '#/components/schemas/SiteState'
        - name: hospital
          in: query
          required: false
          schema:
            type: string
        - name: q
          in: query
          required: false
          schema:
            type: string
            maxLength: 64
      responses:
        '200':
          description: 中心台账
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListStudySitesResponse'
        '401':
          description: 未认证
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/unauthenticated
                title: 未认证
                status: 401
                code: unauthenticated
        '403':
          description: 无此动作权限
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/forbidden-action
                title: 无此动作权限
                status: 403
                code: forbidden-action
        '404':
          description: 资源不存在
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/not-found
                title: 资源不存在
                status: 404
                code: not-found
        '422':
          description: 请求校验未通过
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/validation-failed
                title: 请求校验未通过
                status: 422
                code: validation-failed
        '429':
          description: 请求过于频繁
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/rate-limited
                title: 请求过于频繁
                status: 429
                code: rate-limited
        '500':
          description: 服务内部错误
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/internal
                title: 服务内部错误
                status: 500
                code: internal
    post:
      operationId: createStudySite
      tags:
        - site
      summary: 中心建档
      description: 「已建档 / 合同中心数」的差值 = 合同里写了但还没进系统的中心：它们的成本已经在发生，收入却挂不上号。建档滞后是早期成本失控最不显形的一种。
      x-layer: L1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStudySiteRequest'
      responses:
        '201':
          description: 中心建档
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudySite'
        '401':
          description: 未认证
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/unauthenticated
                title: 未认证
                status: 401
                code: unauthenticated
        '403':
          description: 无此动作权限
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/forbidden-action
                title: 无此动作权限
                status: 403
                code: forbidden-action
        '404':
          description: 资源不存在
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/not-found
                title: 资源不存在
                status: 404
                code: not-found
        '422':
          description: 请求校验未通过
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/validation-failed
                title: 请求校验未通过
                status: 422
                code: validation-failed
        '429':
          description: 请求过于频繁
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/rate-limited
                title: 请求过于频繁
                status: 429
                code: rate-limited
        '500':
          description: 服务内部错误
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/internal
                title: 服务内部错误
                status: 500
                code: internal
  /v1/study-sites/{id}:
    get:
      operationId: getStudySite
      tags:
        - site
      summary: 中心详情
      x-layer: L1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: >-
              ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
      responses:
        '200':
          description: 中心详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudySite'
        '401':
          description: 未认证
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/unauthenticated
                title: 未认证
                status: 401
                code: unauthenticated
        '403':
          description: 无此动作权限
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/forbidden-action
                title: 无此动作权限
                status: 403
                code: forbidden-action
        '404':
          description: 资源不存在
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/not-found
                title: 资源不存在
                status: 404
                code: not-found
        '422':
          description: 请求校验未通过
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/validation-failed
                title: 请求校验未通过
                status: 422
                code: validation-failed
        '429':
          description: 请求过于频繁
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/rate-limited
                title: 请求过于频繁
                status: 429
                code: rate-limited
        '500':
          description: 服务内部错误
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/internal
                title: 服务内部错误
                status: 500
                code: internal
  /v1/study-sites/{id}:advance:
    post:
      operationId: advanceStudySite
      tags:
        - site
      summary: 推进中心阶段
      description: |-
        **本系统 L2 命令层的样板。**

        推进不是给一个字段赋值，而是断言一组前置条件已经成立：
        · 推进到 siv：启动清单的阻塞项必须清零
        · 推进到 closed：七项前置条件全部满足（在组受试者、未关闭质疑与质量事件、药品数量平衡、回收药品已销毁、样本链闭环、补偿已发放且有凭证、结题报告已获批）

        未满足时返回 422 `gate-not-satisfied`，`unmet` 里逐条说明还差什么、去哪里处理 —— 而不是一个被禁用的按钮。
      x-layer: L2
      x-required-action: advance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: >-
              ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
        - name: Idempotency-Key
          in: header
          required: true
          description: 幂等键（uuid）。24 小时内重放同一键返回首次结果。
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdvanceStudySiteRequest'
      responses:
        '200':
          description: 推进中心阶段
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdvanceStudySiteResponse'
        '401':
          description: 未认证
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/unauthenticated
                title: 未认证
                status: 401
                code: unauthenticated
        '403':
          description: 无此动作权限
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/forbidden-action
                title: 无此动作权限
                status: 403
                code: forbidden-action
        '404':
          description: 资源不存在
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/not-found
                title: 资源不存在
                status: 404
                code: not-found
        '409':
          description: 版本冲突：资源已被他人修改
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/conflict-version
                title: 版本冲突：资源已被他人修改
                status: 409
                code: conflict-version
        '422':
          description: 请求校验未通过
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/validation-failed
                title: 请求校验未通过
                status: 422
                code: validation-failed
        '429':
          description: 请求过于频繁
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/rate-limited
                title: 请求过于频繁
                status: 429
                code: rate-limited
        '500':
          description: 服务内部错误
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/internal
                title: 服务内部错误
                status: 500
                code: internal
...
  /v1/study-sites/{id}/gate:
    get:
      operationId: getSiteGate
      tags:
        - site
      summary: 推进闸门预检
      description: 在按钮点下去之前就告诉用户还差什么。前端用它决定按钮是否可用，以及在旁边列出未满足项与「去处理」入口。
      x-layer: L1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: >-
              ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
        - name: to
          in: query
          required: false
          description: 缺省为状态机的下一节点
          schema:
            allOf:
              - $ref: '#/components/schemas/SiteState'
      responses:
        '200':
          description: 推进闸门预检
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteGate'
        '401':
          description: 未认证
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/unauthenticated
                title: 未认证
                status: 401
                code: unauthenticated
        '403':
          description: 无此动作权限
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/forbidden-action
                title: 无此动作权限
                status: 403
                code: forbidden-action
        '404':
          description: 资源不存在
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/not-found
                title: 资源不存在
                status: 404
                code: not-found
        '422':
          description: 请求校验未通过
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/validation-failed
                title: 请求校验未通过
                status: 422
                code: validation-failed
        '429':
          description: 请求过于频繁
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/rate-limited
                title: 请求过于频繁
                status: 429
                code: rate-limited
        '500':
          description: 服务内部错误
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/internal
                title: 服务内部错误
                status: 500
                code: internal
...
components:
  schemas:
    FieldIssue:
      type: object
      properties:
        path:
          type: string
          description: JSON Pointer，如 /hours
        message:
          type: string
      required:
        - path
        - message
      additionalProperties: false
    GateUnmet:
      type: object
      properties:
        code:
          type: string
          description: 如 open-queries / ip-imbalance
        message:
          type: string
          description: 面向人的说明，如「3 条数据质疑未关闭」
        module:
          description: 处理入口所属模块，供前端给出「去处理」跳转
          type: string
      required:
        - code
        - message
      additionalProperties: false
    Problem:
      type: object
      properties:
        type:
          type: string
          description: https://sitedesk.dev/problems/{code}
        title:
          type: string
        status:
          type: integer
          minimum: -9007199254740991
          maximum: 9007199254740991
        code:
          type: string
          enum:
            - validation-failed
            - unauthenticated
            - forbidden-action
            - not-found
            - conflict-version
            - idempotency-key-reused
            - invariant-violated
            - gate-not-satisfied
            - rate-limited
            - internal
        detail:
          type: string
        instance:
          description: 出错的请求路径
          type: string
        traceId:
          type: string
        issues:
          type: array
          items:
            $ref: '#/components/schemas/FieldIssue'
        unmet:
          type: array
          items:
            $ref: '#/components/schemas/GateUnmet'
        invariant:
          type: string
      required:
        - type
        - title
        - status
        - code
      additionalProperties: false
      description: RFC 9457 Problem Details。所有非 2xx 响应都是这个形状。
...
    SideEffect:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/SideEffectType'
        summary:
          type: string
        ref:
          type: string
          format: uuid
          pattern: >-
            ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
        amountCents:
          type: integer
          minimum: 0
          maximum: 1000000000000
          description: 金额，单位：分（整数）
        studySiteId:
          type: string
          format: uuid
          pattern: >-
            ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
      required:
        - type
        - summary
      additionalProperties: false
      description: 命令执行时被连带触发的领域事件。客户端必须忽略不认识的 type。
...
    SiteState:
      type: string
      enum:
        - intake
        - irb_submit
        - irb_approve
        - contract
        - siv
        - enrolling
        - enrolled
        - followup
        - closed
      description: >-
        中心状态机。intake 立项 → irb_submit 伦理递交 → irb_approve 伦理批件 → contract 合同签署 → siv SIV启动 → enrolling 入组中 →
        enrolled 入组完成 → followup 随访中 → closed 中心关闭
    Study:
      type: object
      properties:
        id:
          type: string
          format: uuid
          pattern: >-
            ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
        code:
          type: string
          minLength: 1
          maxLength: 64
          pattern: ^[A-Za-z0-9][A-Za-z0-9._-]*$
        shortName:
          type: string
        sponsorName:
          type: string
        phase:
          type: string
        indication:
          type: string
        plannedSubjects:
          type: integer
          minimum: 0
          exclusiveMinimum: true
          maximum: 9007199254740991
        contractAmountCents:
          x-gated-by: price
          description: 【报价与合同金额】无权限时该字段不出现在响应中（不是 null）
          type: integer
          minimum: 0
          maximum: 1000000000000
        startedOn:
          nullable: true
          type: string
          format: date
          pattern: >-
            ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))$
        endsOn:
          nullable: true
          type: string
          format: date
          pattern: >-
            ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))$
      required:
        - id
        - code
        - shortName
        - sponsorName
        - phase
        - indication
        - plannedSubjects
        - startedOn
        - endsOn
      additionalProperties: false
    StudySite:
      type: object
      properties:
        id:
          type: string
          format: uuid
          pattern: >-
            ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
        code:
          type: string
          minLength: 1
          maxLength: 64
          pattern: ^[A-Za-z0-9][A-Za-z0-9._-]*$
        study:
          type: object
          properties:
            id:
              type: string
              format: uuid
              pattern: >-
                ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
            code:
              type: string
              minLength: 1
              maxLength: 64
              pattern: ^[A-Za-z0-9][A-Za-z0-9._-]*$
            shortName:
              type: string
          required:
            - id
            - code
            - shortName
          additionalProperties: false
        hospital:
          type: string
        dept:
          type: string
        city:
          type: string
        piName:
          type: string
        piAccountId:
          nullable: true
          type: string
          format: uuid
          pattern: >-
            ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
        state:
          $ref: '#/components/schemas/SiteState'
        contracted:
          type: integer
          minimum: 0
          exclusiveMinimum: true
          maximum: 9007199254740991
          description: 合同例数
        unitPriceCents:
          x-gated-by: price
          description: 【报价与合同金额】无权限时该字段不出现在响应中（不是 null）
          type: integer
          minimum: 0
          maximum: 1000000000000
        startupFeeCents:
          x-gated-by: price
          description: 【报价与合同金额】无权限时该字段不出现在响应中（不是 null）
          type: integer
          minimum: 0
          maximum: 1000000000000
        irbApprovedOn:
          nullable: true
          type: string
          format: date
          pattern: >-
            ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))$
        sivOn:
          nullable: true
          type: string
          format: date
          pattern: >-
            ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))$
        sivPlannedOn:
          nullable: true
          type: string
          format: date
          pattern: >-
            ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))$
        fpiOn:
          nullable: true
          description: 首例入组日
          type: string
          format: date
          pattern: >-
            ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))$
      required:
        - id
        - code
        - study
        - hospital
        - dept
        - city
        - piName
        - piAccountId
        - state
        - contracted
        - irbApprovedOn
        - sivOn
        - sivPlannedOn
        - fpiOn
      additionalProperties: false
      description: 项目 × 中心。带 x-gated-by 的字段在无权限时**从响应中消失**，不是返回 null。
    SiteGate:
      type: object
      properties:
        from:
          $ref: '#/components/schemas/SiteState'
        to:
          $ref: '#/components/schemas/SiteState'
        satisfied:
          type: boolean
        unmet:
          type: array
          items:
            $ref: '#/components/schemas/GateUnmet'
      required:
        - from
        - to
        - satisfied
        - unmet
      additionalProperties: false
...
    ListStudiesResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Study'
        nextCursor:
          nullable: true
          description: null 表示已到末页
          type: string
          minLength: 1
          maxLength: 512
      required:
        - items
        - nextCursor
      additionalProperties: false
    ListStudySitesResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/StudySite'
        nextCursor:
          nullable: true
          description: null 表示已到末页
          type: string
          minLength: 1
          maxLength: 512
      required:
        - items
        - nextCursor
      additionalProperties: false
    CreateStudySiteRequest:
      type: object
      properties:
        studyId:
          type: string
          format: uuid
          pattern: >-
            ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
        code:
          type: string
          minLength: 1
          maxLength: 64
        hospital:
          type: string
          minLength: 1
          maxLength: 128
        dept:
          type: string
          minLength: 1
          maxLength: 64
        city:
          type: string
          minLength: 1
          maxLength: 32
        piName:
          type: string
          minLength: 1
          maxLength: 64
        piAccountId:
          nullable: true
          type: string
          format: uuid
          pattern: >-
            ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
        contracted:
          type: integer
          minimum: 0
          exclusiveMinimum: true
          maximum: 9007199254740991
        unitPriceCents:
          type: integer
          minimum: 0
          maximum: 1000000000000
          description: 金额，单位：分（整数）
        startupFeeCents:
          default: 0
          type: integer
          minimum: 0
          maximum: 1000000000000
          description: 金额，单位：分（整数）
        sivPlannedOn:
          nullable: true
          type: string
          format: date
          pattern: >-
            ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))$
      required:
        - studyId
        - code
        - hospital
        - dept
        - city
        - piName
        - contracted
        - unitPriceCents
        - startupFeeCents
      additionalProperties: false
    AdvanceStudySiteRequest:
      type: object
      properties:
        to:
          $ref: '#/components/schemas/SiteState'
        reason:
          type: string
          minLength: 4
          maxLength: 500
          description: 变更原因。会写入审计轨迹；核查时真正被问的就是这一栏。
      required:
        - to
      additionalProperties: false
    AdvanceStudySiteResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/StudySite'
        sideEffects:
          type: array
          items:
            $ref: '#/components/schemas/SideEffect'
          description: 本次命令连带触发的事件。为空数组表示没有连带影响。
      required:
        - data
        - sideEffects
      additionalProperties: false
...
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://sso.example/.well-known/openid-configuration
      description: 内部员工：企业微信 / 飞书 OIDC
    magicLink:
      type: http
      scheme: bearer
      description: 外部方（机构办 / 研究者）：一次性魔法链接换取的短期会话令牌
security:
  - oidc: []
  - magicLink: []